Remove CorrectionWithGradient.eval_dict
It does not work well with vectorized calls:
we would like to vectorize over the values of
the dictionary, but e.g. np.vectorized does not do
that out of the box.

We'll have to revise later in what form we can
reintroduce (something like) eval_dict, if any,
e.g. using kwargs instead of an input dict.